requirements.txt
fastapi
uvicorn
python-multipart
numpy
opencv-python
easyocr
pytesseract
Pillow
langdetect
reportlab
pydantic
SQLAlchemy
aiofiles
requests
ruff
mypy
pytest
pydantic-settings
